+++ src/platform/testing/unit/mocha-setup.js
@@ -13,36 +13,11 @@ import '../../site-wide/moment-setup';
 import ENVIRONMENTS from 'site/constants/environments';
 import * as Sentry from '@sentry/browser';
 import { configure } from '@testing-library/dom';
-import fs from 'fs';
-import path from 'path';
 import chaiAxe from './axe-plugin';
 import { sentryTransport } from './sentry';
 
 const isStressTest = process.env.IS_STRESS_TEST;
-
-const ALLOW_LIST = fs.existsSync(path.resolve(`unit_test_allow_list.json`))
-  ? JSON.parse(fs.readFileSync(path.resolve(`unit_test_allow_list.json`)))
-  : [];
-
-const ALL_SPECS = ALLOW_LIST.map(spec => spec.spec_path);
-const DISALLOWED_SPECS = ALLOW_LIST.filter(
-  spec => spec.allowed === false,
-).map(spec => spec.spec_path.substring(spec.spec_path.indexOf('src')));
-const CHANGED_APPS = process.env.CHANGED_FILES
-  ? process.env.CHANGED_FILES.split(' ').map(filePath =>
-      filePath
-        .split('/')
-        .slice(0, 3)
-        .join('/'),
-    )
-  : [];
-const TESTS_TO_STRESS_TEST = ALL_SPECS.filter(specPath =>
-  CHANGED_APPS.some(filePath => specPath.includes(filePath)),
-);
-
-const core = require('@actions/core');
-
-core.exportVariable('TESTS_TO_STRESS_TEST', TESTS_TO_STRESS_TEST);
+const DISALLOWED_SPECS = process.env.DISALLOWED_TESTS || [];
 Sentry.init({
   autoSessionTracking: false,
   dsn: 'http://one@fake/dsn/0',
